deduplicate Sys_ConsoleInput()
Fixes multiline stdin not working in the SDL client, but working in the
dedicated server.
