+++ sys_unix.c
@@ -3,7 +3,6 @@
 #include <windows.h>
 #include <mmsystem.h>
 #include <io.h>
-#include "conio.h"
 #else
 #include <sys/time.h>
 #include <unistd.h>
@@ -69,66 +68,6 @@ void Sys_Print(const char *text)
 #endif
 	}
 	//fprintf(stdout, "%s", text);
-}
-
-char *Sys_ConsoleInput(void)
-{
-	static char text[MAX_INPUTLINE];
-	static unsigned int len = 0;
-#ifdef WIN32
-	int c;
-
-	// read a line out
-	while (_kbhit ())
-	{
-		c = _getch ();
-		if (c == '\r')
-		{
-			text[len] = '\0';
-			_putch ('\n');
-			len = 0;
-			return text;
-		}
-		if (c == '\b')
-		{
-			if (len)
-			{
-				_putch (c);
-				_putch (' ');
-				_putch (c);
-				len--;
-			}
-			continue;
-		}
-		if (len < sizeof (text) - 1)
-		{
-			_putch (c);
-			text[len] = c;
-			len++;
-		}
-	}
-#else
-	fd_set fdset;
-	struct timeval timeout;
-	FD_ZERO(&fdset);
-	FD_SET(0, &fdset); // stdin
-	timeout.tv_sec = 0;
-	timeout.tv_usec = 0;
-	if (select (1, &fdset, NULL, NULL, &timeout) != -1 && FD_ISSET(0, &fdset))
-	{
-		len = read (0, text, sizeof(text) - 1);
-		if (len >= 1)
-		{
-			// rip off the \n and terminate
-			// div0: WHY? console code can deal with \n just fine
-			// this caused problems with pasting stuff into a terminal window
-			// so, not ripping off the \n, but STILL keeping a NUL terminator
-			text[len] = 0;
-			return text;
-		}
-	}
-#endif
-	return NULL;
 }
 
 char *Sys_GetClipboardData (void)
